Refactoring code 1
