+++ src/actions/nodes.js
@@ -42,19 +42,6 @@ export function toggleNodeHovered(nodeHovered) {
   };
 }
 
-export const TOGGLE_PARAMETERS_HOVERED = 'TOGGLE_PARAMETERS_HOVERED';
-
-/**
- * Update the value of the currently-active hovered node
- * @param {string|null} nodeHovered The node's unique identifier
- */
-export function toggleParametersHovered(hoveredParameters) {
-  return {
-    type: TOGGLE_PARAMETERS_HOVERED,
-    hoveredParameters,
-  };
-}
-
 export const TOGGLE_NODE_DATA_LOADING = 'TOGGLE_NODE_DATA_LOADING';
 
 /**
