+++ src/selectors/nodes.js
@@ -2,7 +2,6 @@ import { createSelector } from 'reselect';
 import { select } from 'd3-selection';
 import { arrayToObject } from '../utils';
 import { getPipelineNodeIDs } from './pipeline';
-import { getEdgeswithParams } from './edges';
 import {
   getNodeDisabled,
   getNodeDisabledTag,
@@ -22,6 +21,9 @@ const getTextLabels = (state) => state.textLabels;
 const getFontLoaded = (state) => state.fontLoaded;
 const getNodeTypeDisabled = (state) => state.nodeType.disabled;
 const getClickedNode = (state) => state.node.clicked;
+const getEdgeIDs = (state) => state.edge.ids;
+const getEdgeSources = (state) => state.edge.sources;
+const getEdgeTargets = (state) => state.edge.targets;
 
 /**
  * Gets a map of nodeIds to graph nodes
@@ -234,16 +236,17 @@ export const getVisibleNodes = createSelector(
       : []
 );
 
-/**
- * Gets a map of nodeIds which have parameters as source
- */
-export const getNodeslinkedtoParams = createSelector(
-  [(state) => state.graph.nodes, getEdgeswithParams, getNodeTypeDisabled],
-  (nodes = [], edgeswithParams, nodeTypeDisabled) => {
-    return nodes.reduce((result, node) => {
-      if (edgeswithParams[node.id] && nodeTypeDisabled.parameters === true)
-        result[node.id] = node;
-      return result;
-    }, {});
+export const getNodeswithInputParams = createSelector(
+  [getGraphNodes, getEdgeIDs, getNodeType, getEdgeSources, getEdgeTargets],
+  (nodes, edgeIDs, nodeType, edgeSources, edgeTargets) => {
+    const nodes_list = {};
+    for (const edgeID of edgeIDs) {
+      const source = edgeSources[edgeID];
+      const target = edgeTargets[edgeID];
+      if (nodeType[source] === 'parameters' && nodeType[target] === 'task') {
+        nodes_list[target] = nodes[target];
+      }
+    }
+    return nodes_list;
   }
 );
